add endpoints for adding review on product and getting products review

+++ domain/src/main/java/org/hotiver/domain/Entity/Review.java
@@ -32,7 +32,7 @@ public class Review {
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
 
-    private Short rating;
+    private Integer rating;
 
     private String comment;
